+++ HISTORY.md
@@ -1,6 +1,6 @@
 # 4.1.5 / 2020-09-17
 
-- Replace @ndhoule/defaults with lodash.assign and ES6 spread syntax
+- Replace @ndhoule/defaults with merging via ES6 spread syntax
 
 # 4.1.4 / 2020-09-16